migrate gpt function to new pattern
