+++ features/gpt3.js
@@ -8,26 +8,27 @@ const configuration = new Configuration({
 const openai = new OpenAIApi(configuration);
 
 const { getOptions } = require('../util/shared-helpers.js');
-const { dangerouslyExecuteJS } = require('../util/dangerously-execute-code.js');
+// const { dangerouslyExecuteJS } = require('../util/dangerously-execute-code.js');
+const {fetchRedditFirstPage} = require('../util/fetch-reddit-first-page.js');
 
 const GPT3_PREFIX = '!!';
 
-const { wolframGetShort } = require('../features/wolfram.js');
+// const { wolframGetShort } = require('../features/wolfram.js');
 
 const ConversationContext = require('../util/contextManager.js');
 
-const plugins = {
-  wolfram: {
-    name: 'wolfram',
-    description: 'Use wolfram alpha to get the best answers possible',
-    usage: 'wolfram',
-    examples: ['wolfram("current weather in New York")','wolfram("how many golf balls would fit inside the earth?")'],
-    function: async (query) => {
-      const response = await wolframGetShort({ content: `!7 ${query}`, author: { username: '', discriminator: ''} });
-      return response;
-    }
-  }
-};
+// const plugins = {
+//   wolfram: {
+//     name: 'wolfram',
+//     description: 'Use wolfram alpha to get the best answers possible',
+//     usage: 'wolfram',
+//     examples: ['wolfram("current weather in New York")','wolfram("how many golf balls would fit inside the earth?")'],
+//     function: async (query) => {
+//       const response = await wolframGetShort({ content: `!7 ${query}`, author: { username: '', discriminator: ''} });
+//       return response;
+//     }
+//   }
+// };
 
 
 const systemMessages = [
@@ -110,14 +111,14 @@ const hints = [
 ];
 
 const functions = [
-  {
-    name: 'restart',
-    description: 'Restart the robot',
-    parameters: {
-      type: 'object',
-      properties: {},
-    },
-  },
+  // {
+  //   name: 'restart',
+  //   description: 'Restart the robot',
+  //   parameters: {
+  //     type: 'object',
+  //     properties: {},
+  //   },
+  // },
   // {
   //   name: 'getWeather',
   //   description: 'Get the weather',
@@ -130,25 +131,35 @@ const functions = [
   //     },
   //   },
   // },
-  {
-    name: 'wolphramAlpha',
-    description: 'Use wolfram alpha to get the best answers possible',
-    parameters: {
-      type: 'object',
-      properties: {
-        query: {
-          type: 'string',
-        },
-      },
-    },
-    function: async (query) => {
-      const response = await wolframGetShort({ content: `!7 ${query}`, author: { username: '', discriminator: ''} });
-      return response;
-    },
-  },
+  // {
+  //   name: 'wolphramAlpha',
+  //   description: 'Use wolfram alpha',
+  //   parameters: {
+  //     type: 'object',
+  //     properties: {
+  //       query: {
+  //         type: 'string',
+  //       },
+  //     },
+  //   },
+  //   function: async (query) => {
+  //     const response = await wolframGetShort({ content: `!7 ${query}`, author: { username: '', discriminator: ''} });
+  //     return response;
+  //   },
+  // },
+  // {
+  //   name: 'reddit',
+  //   description: 'fetch the first page of Reddit',
+  //   parameters: {
+  //     type: 'object',
+  //     properties: {},
+  //   },
+  //   function: fetchRedditFirstPage,
+  // },
 ];
 
-const gpt3 = async (message, restartCB, getWeatherData) => {
+// const gpt3 = async (message, restartCB) => {
+const gpt3 = async (message) => {
   const member = message.member;
   const memberId = member.id;
 
@@ -195,7 +206,7 @@ const gpt3 = async (message, restartCB, getWeatherData) => {
     response = await openai.createChatCompletion({
       model: TEXT_MODEL,
       messages: conversation.getContext(), // maybe pass in channel id here?
-      functions: functions,
+      functions: functions.length ? functions : undefined,
       // temperature: 0.9,
       // max_tokens: 150,
       // top_p: 1,
@@ -218,122 +229,58 @@ const gpt3 = async (message, restartCB, getWeatherData) => {
     // check if GPT wants to call a function
     if (response.data.choices[0].message.function_call) {
       const function_call = response.data.choices[0].message.function_call;
-      const function_name = function_call.name;
-      const function_arguments = function_call.arguments;
-
-      console.log('function_call:', function_call);
-
-      if (function_name === 'restart') {
-        restartCB();
-        return;
-      } else if (function_name === 'getWeather') {
-        // parse arguments as json
-        const parsedArguments = JSON.parse(function_arguments);
-
-        const location = parsedArguments.location;
-        const weatherData = await getWeatherData(location);
-        
-        console.log('weatherData:', weatherData);
-
-        // add the weather data to the context and get a new response
-        conversation.addMessage('function', JSON.stringify(weatherData), message, 'getWeather');
-        response = await openai.createChatCompletion({
-          model: TEXT_MODEL,
-          messages: conversation.getContext(),
-          functions: functions,
-          user: memberId,
-        });
-
-        if (!response) {
-          return 'API Error, no response';
-        }
-
-        console.log('response: ', response);
-
-        if (response.status !== 200) {
-          console.log(response.statusText, response.data);
-          return 'API Error';
-        }
-      } else if (function_name === 'wolphramAlpha') {
-        // parse arguments as json
-        const parsedArguments = JSON.parse(function_arguments);
-
-        const query = parsedArguments.query;
-        const wolframAlphaData = await functions.find(f => f.name === 'wolphramAlpha').function(query);
-        
-        console.log('wolframAlphaData:', wolframAlphaData);
-
-        // add the response data to the context and get a new response
-        conversation.addMessage('function', JSON.stringify(wolframAlphaData), message, 'wolphramAlpha');
-        response = await openai.createChatCompletion({
-          model: TEXT_MODEL,
-          messages: conversation.getContext(),
-          functions: functions,
-          user: memberId,
-        });
-
-        if (!response) {
-          return 'API Error, no response';
-        }
-
-        console.log('response: ', response);
-
-        if (response.status !== 200) {
-          console.log(response.statusText, response.data);
-          return 'API Error';
-        }
+      handleFunctionCall(function_call, message, member, memberId);
+      return `Function call handled: ${function_call.name}, ${function_call.parameters}`;
+    } else {
+      let gptMessage = response.data.choices[0].message.content.trim();
+      console.log('gptMessage:', gptMessage);
+      // messages.push({ role: 'assistant', content: gptMessage });
+  
+      // if first two characters are << and last two characters are >>, then we have code to execute
+      // if (gptMessage.slice(0, 2) === '<<' && gptMessage.slice(-2) === '>>') {
+      //   const code_to_execute = gptMessage.slice(2, -2);
+  
+      //   console.log('Code to execute:', code_to_execute);
+  
+      //   const result = await dangerouslyExecuteJS(code_to_execute);
+      //   gptMessage = await gpt3({ ...message, member: {id: memberId, ...member}, content: `${GPT3_PREFIX} Below is the result of the code execution. Please use it to formulate a response to the user:\n${result}` });
+      // }
+  
+  
+      // check for plugin usage
+      // const regex = /\{\{(\w+)\(\"(.*?)\"\)\}\}/;
+      // const match = regex.exec(gptMessage);
+  
+      // if (match !== null) {
+      //   const pluginName = match[1];
+      //   console.log('Plugin name:', pluginName);
+      //   const pluginQuery = match[2];
+      //   console.log('Plugin query:', pluginQuery);
+  
+      //   // check if plugin exists
+      //   if (plugins[match[1]]) {
+      //     const plugin = plugins[match[1]];
+      //     console.log('Plugin found:', plugin.name);
+      //     const pluginResponse = await plugin.function(pluginQuery);
+      //     console.log('Plugin response:', pluginResponse);
+      //     // messages.push({ role: 'assistant', content: pluginResponse });
+      //     // return `${gptMessage}`;
+      //     gptMessage = await gpt3({...message, member: message.member, content: `${GPT3_PREFIX} Below is the response from the plugin. Please use it to formulate a response to the user:\n${pluginResponse}`} );
+      //   }
+      // } else {
+      //   console.log('No plugin found in the input text.');
+      // }
+  
+      // add the AI's message to the conversation
+      conversation.addMessage('assistant', gptMessage, message);
+  
+      // if gptMessage is longer than 2000 characters, split it into multiple messages, each less than 2000 characters
+      if (gptMessage.length > 2000) {
+        const splitMessages = gptMessage.match(/(.|[\r\n]){1,2000}/g);
+        return splitMessages;
       }
-
+      return `${gptMessage}`;
     }
-
-    let gptMessage = response.data.choices[0].message.content.trim();
-    console.log('gptMessage:', gptMessage);
-    // messages.push({ role: 'assistant', content: gptMessage });
-
-    // if first two characters are << and last two characters are >>, then we have code to execute
-    // if (gptMessage.slice(0, 2) === '<<' && gptMessage.slice(-2) === '>>') {
-    //   const code_to_execute = gptMessage.slice(2, -2);
-
-    //   console.log('Code to execute:', code_to_execute);
-
-    //   const result = await dangerouslyExecuteJS(code_to_execute);
-    //   gptMessage = await gpt3({ ...message, member: {id: memberId, ...member}, content: `${GPT3_PREFIX} Below is the result of the code execution. Please use it to formulate a response to the user:\n${result}` });
-    // }
-
-
-    // check for plugin usage
-    // const regex = /\{\{(\w+)\(\"(.*?)\"\)\}\}/;
-    // const match = regex.exec(gptMessage);
-
-    // if (match !== null) {
-    //   const pluginName = match[1];
-    //   console.log('Plugin name:', pluginName);
-    //   const pluginQuery = match[2];
-    //   console.log('Plugin query:', pluginQuery);
-
-    //   // check if plugin exists
-    //   if (plugins[match[1]]) {
-    //     const plugin = plugins[match[1]];
-    //     console.log('Plugin found:', plugin.name);
-    //     const pluginResponse = await plugin.function(pluginQuery);
-    //     console.log('Plugin response:', pluginResponse);
-    //     // messages.push({ role: 'assistant', content: pluginResponse });
-    //     // return `${gptMessage}`;
-    //     gptMessage = await gpt3({...message, member: message.member, content: `${GPT3_PREFIX} Below is the response from the plugin. Please use it to formulate a response to the user:\n${pluginResponse}`} );
-    //   }
-    // } else {
-    //   console.log('No plugin found in the input text.');
-    // }
-
-    // add the AI's message to the conversation
-    conversation.addMessage('assistant', gptMessage, message);
-
-    // if gptMessage is longer than 2000 characters, split it into multiple messages, each less than 2000 characters
-    if (gptMessage.length > 2000) {
-      const splitMessages = gptMessage.match(/(.|[\r\n]){1,2000}/g);
-      return splitMessages;
-    }
-    return `${gptMessage}`;
 
   } catch (error) {
     if (error.response) {
@@ -346,6 +293,87 @@ const gpt3 = async (message, restartCB, getWeatherData) => {
     }
   }
 
+}
+
+const handleFunctionCall = async (function_call, message, member, memberId) => {
+  
+  console.log('function_call:', function_call);
+  const function_name = function_call.name;
+  const function_arguments = function_call.arguments;
+
+
+  // match function name dynamically
+  if (functions.find(f => f.name === function_name)) {
+    const functionObject = functions.find(f => f.name === function_name);
+
+    console.log('functionObject:', functionObject);
+    const functionResponse = await functionObject.function(function_arguments);
+    
+    console.log('functionResponse:', functionResponse);
+
+    gpt3({ ...message, member: {id: memberId, ...member}, content: functionResponse});
+  } 
+
+  // if (function_name === 'restart') {
+  //   restartCB();
+  //   return;
+  // } else if (function_name === 'getWeather') {
+  //   // parse arguments as json
+  //   const parsedArguments = JSON.parse(function_arguments);
+
+  //   const location = parsedArguments.location;
+  //   const weatherData = await getWeatherData(location);
+    
+  //   console.log('weatherData:', weatherData);
+
+  //   // add the weather data to the context and get a new response
+  //   conversation.addMessage('function', JSON.stringify(weatherData), message, 'getWeather');
+  //   response = await openai.createChatCompletion({
+  //     model: TEXT_MODEL,
+  //     messages: conversation.getContext(),
+  //     functions: functions,
+  //     user: memberId,
+  //   });
+
+  //   if (!response) {
+  //     return 'API Error, no response';
+  //   }
+
+  //   console.log('response: ', response);
+
+  //   if (response.status !== 200) {
+  //     console.log(response.statusText, response.data);
+  //     return 'API Error';
+  //   }
+  // } else if (function_name === 'wolphramAlpha') {
+  //   // parse arguments as json
+  //   const parsedArguments = JSON.parse(function_arguments);
+
+  //   const query = parsedArguments.query;
+  //   const wolframAlphaData = await functions.find(f => f.name === 'wolphramAlpha').function(query);
+    
+  //   console.log('wolframAlphaData:', wolframAlphaData);
+
+  //   // add the response data to the context and get a new response
+  //   conversation.addMessage('function', JSON.stringify(wolframAlphaData), message, 'wolphramAlpha');
+  //   response = await openai.createChatCompletion({
+  //     model: TEXT_MODEL,
+  //     messages: conversation.getContext(),
+  //     functions: functions,
+  //     user: memberId,
+  //   });
+
+  //   if (!response) {
+  //     return 'API Error, no response';
+  //   }
+
+  //   console.log('response: ', response);
+
+  //   if (response.status !== 200) {
+  //     console.log(response.statusText, response.data);
+  //     return 'API Error';
+  //   }
+  // }
 }
 
 module.exports = {
